+++ src/parse.js
@@ -1,115 +1,49 @@
-import snakeCase from 'snake-case'
 import parser from './autoSql'
 import types from './defaultTypes'
+import { detectTypes } from './util'
 
-/* eslint no-param-reassign: ["error", { "props": false }] */
-const detectTypes = autoSql => {
-  const numericTypes = ['uint', 'int', 'float', 'long']
-  const fields = autoSql.fields.map(autoField => {
-    const type = {}
-    if (!autoField.size && numericTypes.includes(autoField.type)) {
-      type.isNumeric = true
-    }
-    if (autoField.size) {
-      type.isArray = true
-    }
-    if (autoField.size && numericTypes.includes(autoField.type)) {
-      type.arrayIsNumeric = true
-    }
-    return { ...autoField, ...type }
-  })
-  return { ...autoSql, fields }
-}
 export default class BED {
   constructor(args = {}) {
     if (args.autoSql) {
       this.autoSql = detectTypes(parser.parse(args.autoSql))
-    } else if (args.type && types[args.type]) {
+    } else if (args.type) {
       this.autoSql = detectTypes(
         types[args.type] || throw new Error('Type not found'),
       )
+    } else {
+      this.autoSql = detectTypes(types.defaultBedSchema)
     }
-  }
-
-  parseLine(line) {
-    const { autoSql } = this
-    if (line.startsWith('track') || line.startsWith('browser'))
-      throw new Error(
-        `Error: track and browser line parsing is not supported, please filter:\n${line}`,
-      )
-
-    const ret = line.split('\t')
-    if (ret.length < 3)
-      throw new Error(
-        'Error: line not tab delimited? (note: specification for BED allows spaces but this is unimplemented)',
-      )
-    if (autoSql) {
-      const [refName, start, end, ...rest] = ret
-      return this.parseBedText(refName, +start, +end, rest, {
-        offset: 3,
-      })
-    }
-    return BED.parseBedDefault(ret)
-  }
-
-  static unescape(s) {
-    return s.replace(/%([0-9A-Fa-f]{2})/g, (match, seq) =>
-      String.fromCharCode(parseInt(seq, 16)),
-    )
-  }
-
-  static featureNames = 'refName start end name score strand thickStart thickEnd itemRgb blockCount blockSizes blockStarts'.split(
-    ' ',
-  )
-
-  static parseBedDefault(ret) {
-    const f = ret.map(a => (a === '.' ? null : a))
-
-    const parsed = {}
-    for (let i = 0; i < BED.featureNames.length; i += 1) {
-      if (f[i] !== null && f[i] !== undefined) {
-        parsed[BED.featureNames[i]] = f[i]
-      }
-    }
-    if (parsed.start !== null) parsed.start = parseInt(parsed.start, 10)
-    if (parsed.end !== null) parsed.end = parseInt(parsed.end, 10)
-    if (parsed.score != null) parsed.score = parseFloat(parsed.score, 10)
-
-    // unescape only the ref columns
-    parsed.refName = BED.unescape(parsed.refName)
-    parsed.strand = { '+': 1, '-': -1 }[parsed.strand] || 0
-
-    return parsed
   }
 
   /*
-   * refName - string: name of chromosome or refseq
-   * start - number: start 0 based half open coordinate
-   * end - number: end 0 based half open coordinate
-   * rest - a list of fields to parse
-   * opts.offset to skip fields from the autoSql that have already been parsed
-   * opts.uniqueId to supply a uniqueId that was not encoded in the BED file itself
+   * parses a line of text as a BED line with the loaded autoSql schema
+   *
+   * @param line - a BED line as tab delimited text or array
+   * @param opts - supply opts.uniqueId
+   * @return a object representing a feature
    */
-  parseBedText(refName, start, end, rest, opts = {}) {
+  parseLine(line, opts = {}) {
     const { autoSql } = this
-    const { offset, uniqueId } = opts
-    if (!autoSql)
-      throw new Error(
-        'no autoSql configured, please supply autoSql or format to BED constructor',
-      )
-    const featureData = { refName, start, end }
+    const { uniqueId } = opts
+    let fields = line
+    if (!Array.isArray(line)) {
+      if (line.startsWith('track') || line.startsWith('browser'))
+        throw new Error(
+          `track and browser line parsing is not supported, please filter:\n${line}`,
+        )
+      fields = line.split('\t')
+    }
+
+    const featureData = {}
     if (uniqueId) featureData.uniqueId = uniqueId
-    const bedColumns = Array.isArray(rest) ? rest : rest.split('\t')
-    // first three columns (chrom,start,end) are not included in bigBed, so use offset to compensate
-    for (let i = offset; i < autoSql.fields.length; i += 1) {
+
+    for (let i = 0; i < autoSql.fields.length; i += 1) {
       const autoField = autoSql.fields[i]
-      let columnVal = bedColumns[i - offset]
+      let columnVal = fields[i]
       const { isNumeric, isArray, arrayIsNumeric, name } = autoField
       if (columnVal === null || columnVal === undefined) {
-        console.warn(`column ${i} does not exist, expected ${name}`)
         break
       }
-
       if (columnVal !== '.') {
         if (isNumeric) {
           const num = Number(columnVal)
@@ -124,11 +58,15 @@ export default class BED {
           if (arrayIsNumeric) columnVal = columnVal.map(str => Number(str))
         }
 
-        featureData[snakeCase(name)] = columnVal
+        featureData[name] = columnVal
       }
     }
 
-    featureData.strand = { '-': -1, '+': 1 }[featureData.strand] || 0
+    if (featureData.chrom) {
+      featureData.chrom = decodeURIComponent(featureData.chrom)
+    }
+    featureData.strand = { '.': 0, '-': -1, '+': 1 }[featureData.strand] || 0
+
     return featureData
   }
 }